+++ src/test/java/pl/hetman/wiktoria/java/app/socialactivitytracker/dao/CredentialsPropertiesTest.java
@@ -1,4 +1,4 @@
-package pl.hetman.wiktoria.java.app.socialactivitytracker.model;
+package pl.hetman.wiktoria.java.app.socialactivitytracker.dao;
 
 import org.junit.jupiter.api.Test;
 import pl.hetman.wiktoria.java.app.socialactivitytracker.dao.CredentialsProperties;
